requirements.txt
hdbscan==0.8.33
hyperopt==0.2.7
jsonlines==4.0.0
matplotlib==3.8.3
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
plotly==5.20.0
scikit_learn==1.4.1.post1
seaborn==0.13.2
sentence_transformers==2.5.1
tqdm==4.66.2
umap_learn==0.5.5
vaderSentiment==3.3.2
vaderSentiment==3.3.2
